# shared/ts/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Autoniix API",
    "version": "2.0.0",
    "description": "Autoniix API contracts (generated from Zod schemas)"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v2/jobs": {
      "get": {
        "summary": "List jobs",
        "parameters": [
          {
            "schema": {
              "type": "object",
              "properties": {
                "page": {
                  "type": "integer",
                  "minimum": 1,
                  "default": 1
                },
                "page_size": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 100,
                  "default": 20
                },
                "cursor": {
                  "type": "string"
                }
              }
            },
            "required": false,
            "name": "pagination",
            "in": "query"
          },
          {
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": ["pending", "running", "completed", "failed", "paused", "stopped"]
              }
            },
            "required": false,
            "name": "status_filter",
            "in": "query"
          },
          {
            "schema": {
              "type": "object",
              "properties": {
                "start": {
                  "type": "string",
                  "format": "date-time"
                },
                "end": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "required": ["start", "end"]
            },
            "required": false,
            "name": "time_range",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "sort_by",
            "in": "query"
          },
          {
            "schema": {
              "type": "boolean",
              "default": false
            },
            "required": false,
            "name": "sort_desc",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Jobs list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jobs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "workspace_id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "channel_id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "status": {
                            "type": "string",
                            "enum": ["pending", "running", "completed", "failed", "paused", "stopped"]
                          },
                          "config": {
                            "type": "object",
                            "properties": {
                              "num_videos": {
                                "type": "integer",
                                "minimum": 1,
                                "maximum": 100
                              },
                              "niche": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 255
                              },
                              "auto_publish": {
                                "type": "boolean",
                                "default": false
                              },
                              "parameters": {
                                "type": "object",
                                "additionalProperties": {
                                  "type": "string"
                                }
                              }
                            },
                            "required": ["num_videos", "niche"]
                          },
                          "progress": {
                            "type": "object",
                            "nullable": true,
                            "properties": {
                              "total_videos": {
                                "type": "integer",
                                "minimum": 0
                              },
                              "completed_videos": {
                                "type": "integer",
                                "minimum": 0
                              },
                              "failed_videos": {
                                "type": "integer",
                                "minimum": 0
                              },
                              "current_stage": {
                                "type": "string"
                              },
                              "progress_percent": {
                                "type": "number",
                                "minimum": 0,
                                "maximum": 100
                              },
                              "started_at": {
                                "type": "string",
                                "nullable": true,
                                "format": "date-time"
                              },
                              "estimated_completion": {
                                "type": "string",
                                "nullable": true,
                                "format": "date-time"
                              }
                            },
                            "required": [
                              "total_videos",
                              "completed_videos",
                              "failed_videos",
                              "current_stage",
                              "progress_percent",
                              "started_at",
                              "estimated_completion"
                            ]
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updated_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "completed_at": {
                            "type": "string",
                            "nullable": true,
                            "format": "date-time"
                          },
                          "error_message": {
                            "type": "string",
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "workspace_id",
                          "channel_id",
                          "status",
                          "config",
                          "progress",
                          "created_at",
                          "updated_at",
                          "completed_at",
                          "error_message"
                        ]
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "page": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "page_size": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "next_cursor": {
                          "type": "string",
                          "nullable": true
                        },
                        "has_more": {
                          "type": "boolean"
                        }
                      },
                      "required": ["total", "page", "page_size", "next_cursor", "has_more"]
                    }
                  },
                  "required": ["jobs", "pagination"]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "channel_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "config": {
                    "type": "object",
                    "properties": {
                      "num_videos": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 100
                      },
                      "niche": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 255
                      },
                      "auto_publish": {
                        "type": "boolean",
                        "default": false
                      },
                      "parameters": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "string"
                        }
                      }
                    },
                    "required": ["num_videos", "niche"]
                  }
                },
                "required": ["channel_id", "config"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "job": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "workspace_id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "channel_id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "status": {
                          "type": "string",
                          "enum": ["pending", "running", "completed", "failed", "paused", "stopped"]
                        },
                        "config": {
                          "type": "object",
                          "properties": {
                            "num_videos": {
                              "type": "integer",
                              "minimum": 1,
                              "maximum": 100
                            },
                            "niche": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 255
                            },
                            "auto_publish": {
                              "type": "boolean",
                              "default": false
                            },
                            "parameters": {
                              "type": "object",
                              "additionalProperties": {
                                "type": "string"
                              }
                            }
                          },
                          "required": ["num_videos", "niche"]
                        },
                        "progress": {
                          "type": "object",
                          "nullable": true,
                          "properties": {
                            "total_videos": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "completed_videos": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "failed_videos": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "current_stage": {
                              "type": "string"
                            },
                            "progress_percent": {
                              "type": "number",
                              "minimum": 0,
                              "maximum": 100
                            },
                            "started_at": {
                              "type": "string",
                              "nullable": true,
                              "format": "date-time"
                            },
                            "estimated_completion": {
                              "type": "string",
                              "nullable": true,
                              "format": "date-time"
                            }
                          },
                          "required": [
                            "total_videos",
                            "completed_videos",
                            "failed_videos",
                            "current_stage",
                            "progress_percent",
                            "started_at",
                            "estimated_completion"
                          ]
                        },
                        "created_at": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updated_at": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "completed_at": {
                          "type": "string",
                          "nullable": true,
                          "format": "date-time"
                        },
                        "error_message": {
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "workspace_id",
                        "channel_id",
                        "status",
                        "config",
                        "progress",
                        "created_at",
                        "updated_at",
                        "completed_at",
                        "error_message"
                      ]
                    }
                  },
                  "required": ["job"]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/jobs/{job_id}/pause": {
      "post": {
        "summary": "Pause job",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "job_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Job paused"
          }
        }
      }
    },
    "/api/v2/jobs/{job_id}/resume": {
      "post": {
        "summary": "Resume job",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "job_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Job resumed"
          }
        }
      }
    },
    "/api/v2/jobs/{job_id}/retry": {
      "post": {
        "summary": "Retry job",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "job_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Job retried"
          }
        }
      }
    }
  }
}